+++ src/dashboardWebView/components/Media/Media.tsx
@@ -16,6 +16,8 @@ import { FolderItem } from './FolderItem';
 import useMedia from '../../hooks/useMedia';
 import { TelemetryEvent } from '../../../constants';
 import { PageLayout } from '../Layout/PageLayout';
+import { parseWinPath } from '../../../helpers/parseWinPath';
+import { join } from 'path';
 
 export interface IMediaProps {}
 
@@ -26,6 +28,25 @@ export const Media: React.FunctionComponent<IMediaProps> = (props: React.PropsWi
   const selectedFolder = useRecoilValue(SelectedMediaFolderAtom);
   const folders = useRecoilValue(MediaFoldersAtom);
   const loading = useRecoilValue(LoadingAtom);
+
+
+  const allFolders = React.useMemo(() => {
+    // Check if content allows page bundle
+    if (viewData && viewData.data && typeof viewData.data.pageBundle !== "undefined" && !viewData.data.pageBundle) {
+      return folders.filter(f => parseWinPath(f).includes(join('/', settings?.staticFolder || '', '/')));
+    }
+
+    return folders;
+  }, [folders, viewData, settings?.staticFolder]);
+
+  const allMedia = React.useMemo(() => {
+    // Check if content allows page bundle
+    if (viewData && viewData.data && typeof viewData.data.pageBundle !== "undefined" && !viewData.data.pageBundle) {
+      return media.filter(m => parseWinPath(m.fsPath).includes(join('/', settings?.staticFolder || '', '/')));
+    }
+
+    return media;
+  }, [media, viewData, settings?.staticFolder]); 
 
   const onDrop = useCallback((acceptedFiles: File[]) => {
     acceptedFiles.forEach((file) => {
@@ -79,7 +100,7 @@ export const Media: React.FunctionComponent<IMediaProps> = (props: React.PropsWi
         }
 
         {
-          (media.length === 0 && folders.length === 0 && !loading) && (
+          (allMedia.length === 0 && folders.length === 0 && !loading) && (
             <div className={`flex items-center justify-center h-full`}>
               <div className={`max-w-xl text-center`}>
                 <FrontMatterIcon className={`text-vulcan-300 dark:text-whisper-800 h-32 mx-auto opacity-90 mb-8`} />
@@ -91,11 +112,11 @@ export const Media: React.FunctionComponent<IMediaProps> = (props: React.PropsWi
         }
 
         {
-          folders && folders.length > 0 && (
+          allFolders && allFolders.length > 0 && (
             <div className={`mb-8`}>
               <List gap={0}>
                 {
-                  folders && folders.map((folder) => (
+                  allFolders.map((folder) => (
                     <FolderItem key={folder} folder={folder} staticFolder={settings?.staticFolder} wsFolder={settings?.wsFolder} />
                   ))
                 }
@@ -106,7 +127,7 @@ export const Media: React.FunctionComponent<IMediaProps> = (props: React.PropsWi
 
         <List>
           {
-            media.map((file) => (
+            allMedia.map((file) => (
               <Item key={file.fsPath} media={file} />
             ))
           }